feat(TryMe): add ability to try thumbifier from about page
The about page now has a `Try Me` section that allows a website user to
enter a URL to a file and try the thumbifier service. The generated
thumbnail is then displayed in a modal for the website user to view.

Additional styling has been added to the `Try Me` section to keep it in
line with the rest of the site styles.

The `Try Me` section relies on two new files:

  * `callback.php` - receives the thumbifed payload and saves to cache
  * `results.php` - queried by the website for the callback results

The `Try Me` thumbified payloads are cached inside the root `cache`
directory and this directory has all items older than 3 minutes deleted
everytime the website is refreshed, so the directory should never fill
up.