feat(TryMe): move api call to server side component
Previously, the `Try Me` functionality worked completely client side by
having the access token injected into the javascript. Whilst this worked
in practice it meant that anyone could scrape the site for the access
token and use the test account to generate thumbnails for free.

The most appropriate use of the access token is server side where it
cannot be misused. It is still worthy having the ability to use the
access token client side just in case a user wants that functionality as
it is a temporary grant to the api so not a long term security issue for
many.