feat(Docker): add docker container for deployment playground
An optional docker container can now be created to quickly get a
deployed version of the website running.

It comes with a simple bash script to aid in building the website
container and also to kill it if required.

To make the code compatible with the container, the array structure has
been updated to be PHP 5.4 compatible.