fix(supported): truncate long mime-types to prevent overflow
If a mime-type is very long then it will break out of the styled
container on the `/supported` page and look unsightly.

The mime-type containers are now styled to truncate long mime-types and
ellipsise the string. This now means that some mime-types may not be
visible so an `alt` and `title` tag have been added to allow mouse hover
to display the full mime-type.