chore(https): update site to reflect api switch to https
The api is now using https so the documentation for the site requires to
reflect this. Also, the `Try Me` code has been updated to ensure it
works with an SSL connection.

At present the `SSL_VERIFYPEER` for `cURL` has been set to false as
there seems to be an issue with the version of PHP and the certificate.
This is not an ideal solution as it is a security flaw, however in this
instance because the service is implicitly known then the security risk
is limited.