chore(debug): add debug helper files to aid users setup
Occasionally, users of the service complain that their callback isn't
being called. This tends to be an issue with something their side.
However, to make it easier for them to debug these issues a couple of
debug pages have been created.

The `debug_callback.php` allows a user to set this as their callback,
which the thumbifier will push the results to. It simply writes out the
results to a text file `debug_callback_results.txt`.

The user can then view their results by visiting
`debug_callback_results.php` via the official website address.